fix: add vs2022 lib & proper version split
