+++ vs-commitizen.vs2022/GitChangesSection.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel;
+using vs_commitizen.vs;
+
+namespace vs_commitizen.vs2015
+{
+    public partial class GitChangeSection : TeamExplorerBaseSection
+    {
+        private void TeamExplorerPageBasePropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+        }
+    }
+}
