+++ vs-commitizen.vs2022/vs-commitizen.vs2022.csproj
@@ -71,6 +71,7 @@
     <Reference Include="UIAutomationProvider" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="GitChangesSection.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
